Flesh out documentation
