+++ Sources/Middleman/Internal/Convenience Extensions/Set.swift
@@ -3,15 +3,12 @@
 //  Middleman
 //
 //  Created by Valentin Walter on 5/2/20.
-//  
-//
-//  Abstract:
-//
 //
 
 import Foundation
 
 extension Set {
+	/// Adds two sets via `Set.union(_:)`.
     static func += <S: Sequence>(lhs: inout Self, rhs: S) where Element == S.Element {
         lhs = lhs.union(rhs)
     }
