+++ Sources/Middleman/Internal/Convenience Extensions/String+prefixed.swift
@@ -4,15 +4,12 @@
 //
 //  Created by Valentin Walter on 4/29/20.
 //  
-//
-//  Abstract:
-//
-//
 
 import Foundation
 
 extension String {
-    /// Prefixes the string with the given prefix it is not already prefixed with it.
+    /// Prefixes the string with the given prefix it is not already prefixed
+	/// with it.
     func prefixed(with prefix: String) -> String {
         if hasPrefix(prefix) {
             return self
@@ -21,7 +18,8 @@ extension String {
         }
     }
 
-    /// Prefixes the string with the given prefix it is not already prefixed with it.
+    /// Prefixes the string with the given prefix it is not already prefixed
+	/// with it.
     mutating func prefix(with prefix: String) {
         self = prefixed(with: prefix)
     }
