+++ Sources/Middleman/Public/Action/Action+run.swift
@@ -1,13 +1,9 @@
 //
-//  Action+Run.swift
+//  Action+run.swift
 //  Middleman
 //
 //  Created by Valentin Walter on 4/15/20.
 //  
-//
-//  Abstract:
-//  The function that opens the action in the current environment with the specified input.
-//
 
 import Foundation
 
@@ -17,16 +13,19 @@ import Cocoa
 import UIKit
 #endif
 
-public extension Action {
-    /// Open the `Endpoint` of this action in the current workspace.
-    /// - Parameter input: The input with which to run this action.
-    /// - Parameter callback: Called when the client answers either
-    /// of the x-callback parameters.
-    fileprivate func _run(on app: App, with input: Input?, then callback: Callback<Output?>?) {
+extension Action {
+	/// This function contains the actual logic of creating the URL, opening it
+	/// with the platform-specific API and registering the action with
+	/// Middleman.
+	/// - Parameters:
+	///   - app: The `App` (url-scheme) with which to run this action.
+	///   - input: The `Action`'s `Input` associated type.
+	///   - callback: A callback with a `Response` to switch over.
+    internal func _run(on app: App, with input: Input?, then callback: Callback<Output?>?) {
         // Convert this action to a URL
         guard let xurl = toXCallbackURL(app: app,
                                         input: input,
-                                        callback: callback != nil),
+                                        hasCallback: callback != nil),
               let url = xurl.url else
         {
             print(
@@ -45,9 +44,9 @@ public extension Action {
         UIApplication.shared.open(url)
         #endif
 
-        // Store callback identified by UUID (will be called
-        // in `Middleman.receive(url:)` with the decoder fit
-        // to create this action's `Ouput`)
+        // Store callback identified by a UUID (will be called in
+		// `Middleman.receive(url:)` with the decoder fit to create this
+		// action's `Ouput`)
         guard let callback = callback else { return }
 		guard Middleman.receiver != nil else {
 			print("""
@@ -86,27 +85,5 @@ public extension Action {
                 callback(.cancel)
             }
         }
-    }
-}
-
-public extension Action {
-	func run(on app: App, with input: Input, then callback: Callback<Output?>?) {
-		_run(on: app, with: input, then: callback)
-	}
-}
-
-public extension Action where Input == Never {
-	func run(on app: App) {
-		_run(on: app, with: nil, then: nil)
-	}
-	
-	func run(on app: App, then callback: Callback<Output?>?) {
-		_run(on: app, with: nil, then: callback)
-	}
-}
-
-public extension Action where Output == Never {
-    func run(on app: App, with input: Input) {
-        _run(on: app, with: input, then: nil)
     }
 }
