+++ Sources/Middleman/Public/App/Receiver+receive.swift
@@ -4,15 +4,10 @@
 //
 //  Created by Valentin Walter on 4/23/20.
 //  
-//
-//  Abstract:
-//
-//
 
 import Foundation
 
 public extension Receiver {
-
     func receive(xurl: ResponseURL) throws {
         // Extract action from actions defined as receivable on `App` protocol
         guard let action = receivingActions.first(where: { $0.path == xurl.path }) else {
@@ -21,5 +16,4 @@ public extension Receiver {
 
         try action.receive(xurl.decoder)
     }
-
 }
