+++ Sources/Middleman/Public/Callback.swift
@@ -4,10 +4,6 @@
 //
 //  Created by Valentin Walter on 4/20/20.
 //  
-//
-//  Abstract:
-//  A closure that accepts a generic `Response`.
-//
 
 import Foundation
